move fn_pipe & fn_pipe_factory to new file
